+++ backend/infrahub/cli/git_agent.py
@@ -9,7 +9,9 @@ import typer
 from prometheus_client import start_http_server
 from rich.logging import RichHandler
 
-import infrahub.config as config
+from infrahub import config
+from infrahub.core.initialization import initialization
+from infrahub.database import get_db
 from infrahub.git import handle_message, initialize_repositories_directory
 from infrahub.git.actions import sync_remote_repositories
 from infrahub.lock import initialize_lock
@@ -22,6 +24,7 @@ from infrahub.message_bus.events import (
 )
 from infrahub.message_bus.operations import execute_message
 from infrahub.services import InfrahubServices
+from infrahub.services.adapters.database.graph_database import GraphDatabase
 from infrahub.services.adapters.message_bus.rabbitmq import RabbitMQMessageBus
 from infrahub_client import InfrahubClient
 
@@ -59,7 +62,11 @@ async def subscribe_rpcs_queue(client: InfrahubClient):
     channel = await connection.channel()
     queue = await channel.declare_queue(f"{config.SETTINGS.broker.namespace}.rpcs")
     exchange = await channel.declare_exchange(f"{config.SETTINGS.broker.namespace}.events", type="topic")
-    service = InfrahubServices(client=client, message_bus=RabbitMQMessageBus(exchange=exchange))
+    driver = await get_db()
+    database = GraphDatabase(driver=driver)
+    service = InfrahubServices(client=client, database=database, message_bus=RabbitMQMessageBus(exchange=exchange))
+    async with service.database.session as session:
+        await initialization(session=session)
     log.info("Waiting for RPC instructions to execute .. ")
     async with queue.iterator() as qiterator:
         message: IncomingMessage
