+++ backend/infrahub/services/__init__.py
@@ -5,12 +5,19 @@ from infrahub.message_bus import InfrahubBaseMessage
 from infrahub.message_bus.messages import ROUTING_KEY_MAP
 from infrahub_client import InfrahubClient
 
+from .adapters.database import InfrahubDatabase
 from .adapters.message_bus import InfrahubMessageBus
 
 
 class InfrahubServices:
-    def __init__(self, client: Optional[InfrahubClient] = None, message_bus: Optional[InfrahubMessageBus] = None):
+    def __init__(
+        self,
+        client: Optional[InfrahubClient] = None,
+        database: Optional[InfrahubDatabase] = None,
+        message_bus: Optional[InfrahubMessageBus] = None,
+    ):
         self._client = client
+        self.database = database or InfrahubDatabase()
         self.message_bus = message_bus or InfrahubMessageBus()
 
     @property
