+++ backend/infrahub/services/adapters/database/__init__.py
@@ -1,0 +1,9 @@
+from neo4j import AsyncSession
+
+
+class InfrahubDatabase:
+    """Base class for database access"""
+
+    @property
+    def session(self) -> AsyncSession:
+        raise NotImplementedError()
